Move render-embedded-layout
