+++ plugins/woocommerce/client/admin/client/embedded-body-layout/index.ts
@@ -1,1 +1,2 @@
 export * from './embedded-body-layout';
+export { renderEmbeddedLayout } from './render-embedded-layout';
